Save LamdaLRScheduler
In Python3, pickle doesn't save lamda functions. In the LRScheduler,
we will get an error thrown if we try to save the model with a
linear LRScheduler.

Following from this [SO answer](https://stackoverflow.com/a/52799649/10024149),
we update the code to be able to save the LabelModel with a Linear LR
Schedule.